♻️ Refactor : hooks 분리 및 타입 래핑
