+++ src/features/record/recordMain/types/index.ts
@@ -1,14 +1,12 @@
-export interface RecordOverviewResponse {
-  localDateTime: string;
-  responseCode: number;
-  statusCode: string;
-  message: string;
-  data: {
-    title: string;
-    coverLetters: {
-      question: string;
-      answer: string;
-    }[];
-    imageUrl: string;
-  };
+import { ResponseTypes } from '@/shared/api/types';
+
+export interface RecordOverview {
+  title: string;
+  coverLetters: {
+    question: string;
+    answer: string;
+  }[];
+  imageUrl: string;
 }
+
+export type RecordOverviewResponse = ResponseTypes<RecordOverview>;
